Refactor: 레포지토리 메서드 정리
