+++ src/main/java/com/jigumulmi/admin/place/AdminCustomPlaceRepository.java
@@ -2,13 +2,18 @@ package com.jigumulmi.admin.place;
 
 
 import static com.jigumulmi.config.querydsl.Utils.getOrderSpecifier;
+import static com.jigumulmi.config.querydsl.Utils.nullSafeBuilder;
 import static com.jigumulmi.place.domain.QPlace.place;
+import static com.jigumulmi.place.domain.QPlaceCategoryMapping.placeCategoryMapping;
+import static com.jigumulmi.place.domain.QSubwayStationPlace.subwayStationPlace;
 
 import com.jigumulmi.admin.place.dto.request.AdminGetPlaceListRequestDto;
 import com.jigumulmi.place.domain.Place;
-import com.jigumulmi.place.repository.CustomPlaceRepository;
+import com.jigumulmi.place.vo.PlaceCategoryGroup;
+import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
+import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
@@ -24,8 +29,6 @@ public class AdminCustomPlaceRepository {
 
     private final JPAQueryFactory queryFactory;
 
-    private final CustomPlaceRepository customPlaceRepository;
-
     public Page<Place> getPlaceList(Pageable pageable, AdminGetPlaceListRequestDto requestDto) {
         List<Place> content = queryFactory
             .selectFrom(place)
@@ -38,8 +41,7 @@ public class AdminCustomPlaceRepository {
         JPAQuery<Long> totalCountQuery = queryFactory
             .select(place.count())
             .from(place)
-            .where(placeCondition(requestDto)
-            );
+            .where(placeCondition(requestDto));
 
         return PageableExecutionUtils.getPage(content, pageable, totalCountQuery::fetchOne);
     }
@@ -47,9 +49,39 @@ public class AdminCustomPlaceRepository {
     public BooleanExpression placeCondition(AdminGetPlaceListRequestDto requestDto) {
         return (
             place.isFromAdmin.eq(requestDto.getIsFromAdmin())
-                .and(customPlaceRepository.subwayStationCondition(requestDto.getSubwayStationId()))
-                .and(customPlaceRepository.categoryGroupCondition(requestDto.getCategoryGroup()))
-                .and(customPlaceRepository.placeNameContains(requestDto.getPlaceName()))
+                .and(subwayStationCondition(requestDto.getSubwayStationId()))
+                .and(categoryGroupCondition(requestDto.getCategoryGroup()))
+                .and(placeNameContains(requestDto.getPlaceName()))
         );
+    }
+
+    public BooleanExpression subwayStationCondition(Long subwayStationId) {
+        if (subwayStationId == null) {
+            return null;
+        } else {
+            return place.id.in(
+                JPAExpressions
+                    .select(subwayStationPlace.place.id)
+                    .from(subwayStationPlace)
+                    .where(subwayStationPlace.subwayStation.id.eq(subwayStationId))
+            );
+        }
+    }
+
+    public BooleanBuilder placeNameContains(String name) {
+        return nullSafeBuilder(() -> place.name.contains(name));
+    }
+
+    public BooleanExpression categoryGroupCondition(PlaceCategoryGroup categoryGroup) {
+        if (categoryGroup == null) {
+            return null;
+        } else {
+            return place.id.in(
+                JPAExpressions
+                    .select(placeCategoryMapping.place.id)
+                    .from(placeCategoryMapping)
+                    .where(placeCategoryMapping.categoryGroup.eq(categoryGroup))
+            );
+        }
     }
 }
